Fix namespace and configuration for getserviceid

+++ configuration/configuration_repository.go
@@ -100,9 +100,9 @@ func (s *mainRepository) getServiceId(serviceName string) (int, error) {
 	if err != nil {
 		return 0, err
 	}
-	res, ok := ret["id"].(int)
+	res, ok := ret["id"].(int64)
 	if !ok {
 		return 0, helper.NewHttpError(http.StatusNotFound, "Not Found")
 	}
-	return res, nil
+	return int(res), nil
 }
